Misc (re #1370): minor fix compile warning on VS2005 about comparison between int & unsigned int.
git-svn-id: https://svn.pjsip.org/repos/pjproject/trunk@3790 74dad513-b988-da41-8d7b-12977e46ad98

+++ pjsip/src/pjsua-lib/pjsua_vid.c
@@ -2074,7 +2074,7 @@ PJ_DEF(pj_bool_t) pjsua_call_vid_stream_is_running( pjsua_call_id call_id,
     }
 
     call = &pjsua_var.calls[call_id];
-    PJ_ASSERT_RETURN(med_idx >= 0 && med_idx < call->med_cnt, PJ_EINVAL);
+    PJ_ASSERT_RETURN(med_idx >= 0 && med_idx < (int)call->med_cnt, PJ_EINVAL);
 
     call_med = &call->media[med_idx];
